+++ app/detector.py
@@ -6,7 +6,7 @@ import io
 
 class HoldDetector:
     def __init__(self):
-        model_path = Path(__file__).parent / "weights" / "Ndatayolov5.pt"
+        model_path = Path(__file__).parent.parent / "model" / "weights" / "Ndatayolov5.pt"
         self.model = YOLO(str(model_path))
 
     async def detect(self, image_file) -> dict:
